% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_weighted_average.R
\name{best_weighted_average}
\alias{best_weighted_average}
\title{Find best weights for each model in an ensemble}
\usage{
best_weighted_average(submodels, n_weight)
}
\arguments{
\item{submodels}{Modeltable with models to average}

\item{n_weight}{Number of different weights per weight. Example: n_weight = 3 means expand_grid(w1 = 0:3, w2 = 0:3) for two models}
}
\description{
Find best weights for each model in an ensemble
}
