% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_covid.R
\name{create_covid}
\alias{create_covid}
\title{Creates dummy variable for Covid}
\usage{
create_covid(data, fjolda_mork, horizon, freq)
}
\arguments{
\item{data}{Data frame with one date column named date and one covid column named covid}

\item{fjolda_mork}{Threshold for daily covid cases for the dummy to be 1}
}
\description{
Creates dummy variable for Covid
}
