% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ml_data.R
\name{data_prep_func}
\alias{data_prep_func}
\title{Prepares data for modeling}
\usage{
data_prep_func(
  data,
  outcome_var,
  negative_to_zero = FALSE,
  fix_gap_size = FALSE,
  max_gap_size = 52,
  trailing_zero = FALSE,
  transformation = "none",
  use_holidays = FALSE,
  holidays_to_use_1,
  holidays_to_use_2,
  use_seasonal_lag = TRUE,
  use_covid = FALSE,
  covid_data,
  horizon = 12,
  clean = FALSE,
  drop_na = TRUE,
  use_holiday_to_clean = FALSE,
  holiday_for_clean,
  use_abc_category = FALSE,
  pacf_threshold = 0.2,
  no_fourier_terms = 5,
  fourier_k = 5,
  slidify_period = c(4, 8),
  use_own_fourier = FALSE,
  fourier_terms,
  recursive_data = FALSE,
  no_recursive_lag,
  xreg = NULL,
  remove_one_obs_train = FALSE
)
}
\arguments{
\item{data}{Data frame with data on the right date format e.g. daily, weekly, monthly and with column named 'id'}

\item{outcome_var}{Name of the outcome variable. The function will change the outcome variable name to 'outcome'}

\item{negative_to_zero}{Recodes negative values as zero, defaults to TRUE}

\item{max_gap_size}{The maximum length that the outcome can be zero. If the interval is larger than max_gap_size then only use data after the interval}

\item{trailing_zero}{Extends all time series, back and forward, so they will be the same length. Defaults to FALSE}

\item{transformation}{Should the series be transformed, e.g. log or log1p. Defaults to none}

\item{use_holidays}{Should national holidays be included. As of now this has to be a dataframe supplied by the user to the holidays_to_use argument}

\item{holidays_to_use_1}{Data frame of dummy holidays. Outcome of the create_holiday() function: fridagar_tbl}

\item{holidays_to_use_2}{Data frame of holidays, one variable.}

\item{use_seasonal_lag}{Should lag of outcome variable, equal to the seasonality, be used. Defaults to TRUE.}

\item{use_covid}{Should information on covid be used. Data frame has to be supplied by user to the covid_data argument}

\item{covid_data}{Data frame to capture the effect of covid. Has to include one date column and one covid column}

\item{horizon}{The forecast horizon}

\item{clean}{Should the data be cleand for outliers. Defaults to FALSE}

\item{drop_na}{When creating data_prepared_tbl, should NA's be dropped. Defaults to TRUE}

\item{use_holiday_to_clean}{Uses fridagar_one_var from the create_holiday() function to revert series to original value if cleand}

\item{pacf_threshold}{Threshold for where to cut the PACF to choose terms for the fourier calculation}

\item{no_fourier_terms}{Number of fourier terms, defultas to 5}

\item{fourier_k}{The fourier term order, defaults to 5}

\item{slidify_period}{The window size, defaults to c(4, 8)}

\item{use_own_fourier}{Should you use your own fourier terms? Defaults to FALSE}

\item{fourier_terms}{The fourier terms to include.}

\item{recursive_data}{Should the data be prepared for a recursive forecasting. Defaults to FALSE.}

\item{no_recursive_lag}{The number of lags to be.}

\item{xreg}{External regressors to add}

\item{remove_one_obs_train}{Should id's with one observation in the TRAINING set be removed. Defaults to FALSE.}
}
\value{
List with data_prepared, future_data, train_data, splits and horizon
}
\description{
Prepares data for modeling
}
