% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_one_ts.R
\name{data_prep_single_ts}
\alias{data_prep_single_ts}
\title{A function to prepare a single time series for forecasting algorithms}
\usage{
data_prep_single_ts(
  data,
  dep_var,
  date_group = "day",
  clean_outliers = TRUE,
  drop_na = TRUE,
  prepros = "log",
  spline_degree = 3,
  fourier_k = 5,
  horizon = 30,
  use_xreg = FALSE,
  zero_out_fridagar = today() + years(1),
  fridagar = FALSE
)
}
\arguments{
\item{data}{A data frame that contains only two variables, date and the outcome}

\item{dep_var}{Unquoted name of the dependent variable}

\item{date_group}{Frequency of the data, defaults to day}

\item{drop_na}{When creating data_prepared_tbl, should NA's be dropped. Defaults to TRUE}

\item{prepros}{The preprocessing of the outcome variable. Defaults to log}

\item{spline_degree}{The degree of the natural spline, defaults to 3}

\item{fourier_k}{Number or orders to include for each sine/cosine}

\item{use_xreg}{Are you using external regressors, defaults to FALSE}

\item{zero_out_fridagar}{Should holidays be set to zero for the forecast horizon due to e.g. covid, defaults to 1 year into the future}

\item{fridagar}{Do you want to use holidays, defaults to FALSE}

\item{horzon}{The forecasting horizon, defaults to 30}
}
\description{
data_prep_single_ts() is function that allows you to create machine ready data
}
