% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_nbeats.R
\name{finalize_nbeats}
\alias{finalize_nbeats}
\title{Function to select best N-BEATS based on chosen metric}
\usage{
finalize_nbeats(accuracy_table, id, freq, recipe, splits, metric = "rmse")
}
\arguments{
\item{accuracy_table}{Table from tune_nbeats()}

\item{id}{A quoted column name that tracks the GluonTS FieldName "item_id"}

\item{freq}{A pandas timeseries frequency such as "5min" for 5-minutes or "D" for daily.}

\item{recipe}{GluonTS recipe}

\item{metric}{The metric used to choose the best model}

\item{split}{objects}
}
\description{
Function to select best N-BEATS based on chosen metric
}
