% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isl_holidays.R
\docType{data}
\name{isl_holidays}
\alias{isl_holidays}
\title{Holidays and other events in Iceland}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1632 rows and 2 columns.
}
\usage{
data(isl_holidays)
}
\description{
Data from http://dagarnir.is/
}
\keyword{datasets}
