% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_models_wflw.R
\name{ml_models_wflw}
\alias{ml_models_wflw}
\title{A function to train simple univariate algorithms}
\usage{
ml_models_wflw(
  recipe_base,
  recipe_spline,
  recipe_lag,
  recipe_both,
  data = training(splits),
  use_models = c("glmnet", "mars", "svm_poly", "svm_rbf", "knn", "rf", "xgboost",
    "cubist", "nnet", "nnetar", "arima_boost", "prophet_boost"),
  skip_models = NULL
)
}
\arguments{
\item{recipe_base}{Recipe for the modeltime algorithms and NNETAR}

\item{recipe_spline}{Recipe with natural spliens}

\item{recipe_lag}{Recipe with lags}

\item{recipe_both}{Recipe with splines and lags}

\item{data}{Data to use}

\item{use_models}{Models to use}

\item{skip_models}{Character vector of models to skip, defaults to NULL}
}
\description{
train_simple_models() is a function that allows you to train univariate time series models
}
