% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepared_energy.R
\name{prepare_energy_data}
\alias{prepare_energy_data}
\title{Prepares data for energy forecasting}
\usage{
prepare_energy_data(
  data,
  outcome_var,
  horizon,
  xreg_tbl,
  slidify_period,
  transformation = "none",
  fourier_periods = NULL,
  drop_na = TRUE,
  use_holidays = FALSE,
  holidays_tbl,
  pacf_threshold = 0.2,
  no_fourier_terms = 10,
  fourier_k = 3,
  use_horizon_lag = TRUE
)
}
\arguments{
\item{data}{Data frame to be used. Should contain id column, added if not present}

\item{outcome_var}{Name of the outcome variable. It's renamed as outcome}

\item{horizon}{The forecast horizon denoted as interger value}

\item{xreg_tbl}{Data frame with external regressors}

\item{slidify_period}{Sliding periods used in tk_augment_slidify. If not given c(1/4, 1/3, 1/2, 1) * horizon is used}

\item{transformation}{Transformation to be applied. Defaults to none. Possible values: log, log1p}

\item{pacf_threshold}{Threshold to determine which pacf lags to when creating fourier_period. Defaults to 0.2}

\item{no_fourier_terms}{Number of fourier periods to include. Defaults to 10}

\item{fourier_k}{Maximum order(s) of fourier terms}

\item{use_horizon_lag}{Should lag of outcome variable equal to the horizon be used? Defaults to TRUE}
}
\description{
Prepares data for energy forecasting
}
