% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_ml_models.R
\name{stack_ml_models}
\alias{stack_ml_models}
\title{Creates super-learner by stacking differnt ML models}
\usage{
stack_ml_models(
  submodels_resamples,
  kfold,
  grid,
  models,
  parallel = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{submodels_resamples}{A modeltime table (mdl_time_tbl) object with a column containing resample results. Created with modeltime_fit_resamples() from the modeltime.resample package}

\item{kfold}{Number of folds in K-fold cross validation}

\item{grid}{Size of grid to tune the meta-learner. Gets passed to tune::tune_grid()}

\item{models}{Which models to try for the meta-learner. Possible choices: glmnet, rf, nnet, xgboost, cubist and svm}

\item{parallel}{Should the calculations be run in parallel. Defaults to TRUE}

\item{verbose}{A logical for logging results as they are generated. Defaults to TRUE}
}
\value{
A list of length two with ensemble accuracy and a modeltime_table to be used for forecasting
}
\description{
Creates super-learner by stacking differnt ML models
}
