% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumo_forecast_per_id.R
\name{sumo_forecast_per_id}
\alias{sumo_forecast_per_id}
\title{Assign a different model to each time series}
\usage{
sumo_forecast_per_id(
  model_tbl,
  splits,
  future_data,
  data_prepared,
  transformation
)
}
\arguments{
\item{model_tbl}{A modeltime_table that has been refitted to the whole dataset}

\item{future_data}{Data frame for the future period i.e. a future frame}

\item{data_prepared}{Dataset used for training and testing (data_prepared_tbl)}

\item{A}{split object}
}
\description{
Assign a different model to each time series
}
