% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_simple_models.R
\name{train_simple_models}
\alias{train_simple_models}
\title{A function to train simple univariate algorithms}
\usage{
train_simple_models(data = training(splits), seasonality = "daily")
}
\arguments{
\item{data}{Training data (training(splits)) passed to each model}

\item{seasonality}{Controls the seasonality}
}
\description{
train_simple_models() is a function that allows you to train univariate time series models
}
