% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_nbeats.R
\name{tune_nbeats}
\alias{tune_nbeats}
\title{Function to tune N-BEATS}
\usage{
tune_nbeats(
  id,
  freq,
  recipe,
  splits,
  horizon,
  length,
  cv_slice_limit,
  assess = "12 weeks",
  skip = "4 weeks",
  initial = "12 months",
  epochs = NULL,
  lookback = NULL,
  bagging_size = NULL,
  learn_rate = NULL,
  loss_function = NULL,
  scale = NULL,
  batch_size = NULL
)
}
\arguments{
\item{id}{A quoted column name that tracks the GluonTS FieldName "item_id"}

\item{freq}{A pandas timeseries frequency such as "5min" for 5-minutes or "D" for daily.}

\item{recipe}{A gluonts recipe}

\item{horizon}{The forecast horizon}

\item{length}{The number of distinct hyperparameter for each tunable parameter except loss_function which is set to MASE}

\item{cv_slice_limit}{How many slice/folds in the tsCV}

\item{assess}{The number of samples used for each assessment resample}

\item{skip}{A integer indicating how many (if any) additional resamples to skip to thin the total amount of data points in the analysis resample.}

\item{initial}{The number of samples used for analysis/modeling in the initial resample.}

\item{epochs}{Number of epochs}

\item{lookback}{Lookback length. If NULL, will be randomly chosen}

\item{bagging_size}{The number of models that share the parameter combination of 'context_length' and 'loss_function'.}

\item{learn_rate}{Learning rate}

\item{loss_function}{Any of MASE, MAPE sMAPE. Defaults to MASE when loss_function = NULL}

\item{scale}{Scales numeric data by id group using mean = 0, standard deviation = 1 transformation.}

\item{batch_size}{Number of examples in each batch}
}
\description{
Function to tune N-BEATS
}
