% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflw_creator.R
\name{wflw_creator}
\alias{wflw_creator}
\title{A helper function to create workflow passed to the ml_tune() function}
\usage{
wflw_creator(
  model_spec,
  ml_recipe,
  resamples_kfold,
  grid_size = grid_size,
  parallel_type
)
}
\arguments{
\item{model_spec}{A model specification created with the Parsnip or Modeltime package}

\item{ml_recipe}{Recipe for the models}

\item{resamples_kfold}{Resamples used for tuning parameters}

\item{grid_size}{The size of the grid of parameters}
}
\description{
wflw_creater() is a function that creates workflow
}
